+++ FoodOrderApp/FoodOrderApp/ResponseModel/OrderCompleteRM.cs
@@ -8,8 +8,8 @@ namespace FoodOrderApp.ResponseModel
     public class OrderCompleteRM
     {
         public double OrderTotal { get; set; }
-        public long CustomerId { get; set; }
-        public long MenuItemId { get; set; }
-        public long PaymentTypeId { get; set; }
+        public int CustomerId { get; set; }
+        public int MenuItemId { get; set; }
+        public int PaymentTypeId { get; set; }
     }
 }
